requirements.txt
ib_insync
time
pandas
send_email
ibapi.client 
ibapi.wrapper
threading 
datetime
warnings